+++ ppi/constants.py
@@ -1,0 +1,8 @@
+"""Constant values etc."""
+
+EXIT_ERROR: int = 1
+EXIT_SUCCESS: int = 0
+ENCODING: str = "utf-8"
+LANG_CODES: dict = {
+    "FINNISH": "fi_FI.UTF-8"
+}