+++ ppi/parsing.py
@@ -1,153 +1,178 @@
-"""
-parsing.py: Command line argument parsing.
-Author: Niklas Larsson
-Date: September 10, 2021
-"""
-
+"""Command line argument parsing."""
 
 import sys
 
-from ppi.interface import desc_text
-from ppi.interface import errors
-from ppi.interface import file_operation
-from ppi.interface import git_operation
-from ppi.interface import help_text
-from ppi.interface import success_text
-from ppi.interface import usage_text
-
-from ppi.static import exit_codes
+from ppi import constants
+from ppi import errors
+from ppi import texts
 
 
 class ArgParser:
-    """Class to parse cmd line args with."""
+    """Class for parsing command line arguments."""
 
-    def __init__(self, args: list, name: str, version: str, lang: str) -> None:
-        self.args: list = args
-        self.name: str = name
-        self.version: str = version
-        self.lang: str = lang
-        self.prname: str = None
+    def __init__(self, argv: list, program: str, language: str) -> None:
+        """Initial values."""
+        self.argv: list = argv
+        self.program: str = program
+        self.language: str = language
 
-        # Different argument groups.
-        self.invalid_args: list = None
-        self.opts_long: object
-        self.opts_short: object
-        self.pos_args: object
+        self.options: dict = {
+            "long": None,
+            "short": None
+        }
 
-        # Option switches.
-        self.help_on: bool = False
-        self.version_on: bool = False
-        self.quiet_on: bool = False
-        self.ghrepo_on: bool = False
+        self.arguments: dict = {
+            "invalid": None,
+            "positional": None
+        }
 
-    def __repr__(self) -> str:
-        return f"ArgParser(args={self.args!r})"
+        # Option switches
+        self._project: str = ""
+        self._help: bool = False
+        self._version: bool = False
+        self._quiet: bool = False
+        self._git: bool = False
 
-    def __str__(self) -> str:
-        return f"Args: {self.args[1:]}"
+    @property
+    def project(self) -> str:
+        """Gets project's name."""
+        return self._project
+
+    @project.setter
+    def project(self, value: str) -> None:
+        """Sets project's name."""
+        if isinstance(value, str):
+            self._project = value
+
+    @property
+    def help(self) -> bool:
+        """Checks if -h or --help is requested."""
+        return self._help
+
+    @help.setter
+    def help(self, value: bool) -> None:
+        """Sets self._help."""
+        if value in {True, False}:
+            self._help = value
+
+    @property
+    def version(self) -> bool:
+        """Checks if -V or --version is requested."""
+        return self._version
+
+    @version.setter
+    def version(self, value: bool) -> None:
+        """Sets self._version."""
+        if value in {True, False}:
+            self._version = value
+
+    @property
+    def quiet(self) -> bool:
+        """Checks if -q or --quiet is requested."""
+        return self._quiet
+
+    @quiet.setter
+    def quiet(self, value: bool) -> None:
+        """Sets self._quiet."""
+        if value in {True, False}:
+            self._quiet = value
+
+    @property
+    def git(self) -> bool:
+        """Checks if -i or --git-init is requested."""
+        return self._git
+
+    @git.setter
+    def git(self, value: bool) -> None:
+        """Sets self._git."""
+        if value in {True, False}:
+            self._git = value
 
     def _startswith_hyphens(self, arg: str, count: int) -> bool:
-        """Check if `arg` starts with `count` amount of "-"."""
-        return arg[0:count] == "-"*count and arg[count] != "-"
+        """Checks if arg starts with count amount of "-"."""
+        return arg[0:count] == "-" * count and arg[count] != "-"
 
     def _sort_args_long(self) -> None:
-        self.opts_long = (
-            arg for arg in self.args if self._startswith_hyphens(arg, 2)
+        """Gathers all '--' prefixed options from sys.argv into one group."""
+        self.options["long"] = (
+            arg for arg in self.argv if self._startswith_hyphens(arg, 2)
         )
 
     def _sort_args_short(self) -> None:
-        self.opts_short = (
-            arg for arg in self.args if self._startswith_hyphens(arg, 1)
+        """Gathers all '-' prefixed options from sys.argv into one group."""
+        self.options["short"] = (
+            arg for arg in self.argv if self._startswith_hyphens(arg, 1)
         )
 
     def _sort_args_pos(self) -> None:
-        self.pos_args = (
-            arg for index, arg in enumerate(self.args)
+        """
+        Gathers all arguments with no prefixing
+        from sys.argv into one group.
+        """
+        self.arguments["positional"] = (
+            arg for index, arg in enumerate(self.argv)
             if not arg.startswith("-") and index != 0
         )
 
     def _sort_args(self) -> None:
+        """Sorts all argument types."""
         self._sort_args_long()
         self._sort_args_short()
         self._sort_args_pos()
 
     def _parse_args_inv(self) -> None:
-        if self.invalid_args is not None:
-            for arg in self.invalid_args:
-                errors.invargerror(self.lang, self.name, arg)
-            sys.exit(exit_codes.ERROR)
+        """Prints error for each invalid argument."""
+        if self.arguments["invalid"] is not None:
+            for arg in self.arguments["invalid"]:
+                errors.InvalidArgumentError(self.program, self.language).throw_error(arg)
+            sys.exit(constants.EXIT_ERROR)
 
     def _parse_args_pos(self) -> None:
+        """Stores the first non-prefixed argument from sys.argv."""
         # Grab the first non-flag -argument and ignore the rest.
-        for arg in self.pos_args:
-            self.prname = arg
+        for arg in self.arguments["positional"]:
+            self.project = arg
             break
 
     def _parse_args_short(self) -> None:
-        for arg in self.opts_short:
+        """Evaluates each '-' prefixed option."""
+        for arg in self.options["short"]:
             for index, letter in enumerate(arg):
                 if index == 0:
                     continue  # Skip the "-" prefix.
                 if letter == "V":
-                    self.version_on = True
+                    self.version = True
                 elif letter == "h":
-                    self.help_on = True
+                    self.help = True
                 elif letter == "q":
-                    self.quiet_on = True
+                    self.quiet = True
                 elif letter == "i":
-                    self.ghrepo_on = True
+                    self.git = True
                 else:
-                    if self.invalid_args is None:
-                        self.invalid_args = []
-                    self.invalid_args.append("-{}".format(letter))
+                    if self.arguments["invalid"] is None:
+                        self.arguments["invalid"] = []
+                    self.arguments["invalid"].append(f"-{letter}")
 
     def _parse_args_long(self) -> None:
-        for index, arg in enumerate(self.opts_long):
+        """Evaluates each '--' prefixed option."""
+        for index, arg in enumerate(self.options["long"]):
             if arg == "--version":
-                self.version_on = True
+                self.version = True
             elif arg == "--help":
-                self.help_on = True
+                self.help = True
             elif arg == "--quiet":
-                self.quiet_on = True
+                self.quiet = True
             elif arg == "--git-init":
-                self.ghrepo_on = True
+                self.git = True
             else:
-                if self.invalid_args is None:
-                    self.invalid_args = []
-                self.invalid_args.append(arg)
-
-    def _check_if_args(self) -> None:
-        if len(self.args) == 1:
-            desc_text.show(self.name, self.version, self.lang)
-            usage_text.show(self.name, self.version, self.lang)
-            sys.exit(exit_codes.ERROR)
+                if self.arguments["invalid"] is None:
+                    self.arguments["invalid"] = []
+                self.arguments["invalid"].append(arg)
 
     def parse_args(self) -> None:
         """Parse args and execute actions according to the given options."""
-        self._check_if_args()
         self._sort_args()
         self._parse_args_pos()
         self._parse_args_short()
         self._parse_args_long()
         self._parse_args_inv()
-        self._exec_actions()
-
-    def _exec_actions(self) -> None:
-        if self.help_on:
-            desc_text.show(self.name, self.version, self.lang)
-            usage_text.show(self.name, self.version, self.lang)
-            help_text.show(self.name, self.lang)
-            sys.exit(exit_codes.SUCCESS)
-        if self.version_on and not self.help_on:
-            desc_text.show(self.name, self.version, self.lang)
-            sys.exit(exit_codes.SUCCESS)
-        if self.prname is not None:
-            file_operation.create(self.lang, self.name, self.prname)
-            if self.ghrepo_on:
-                git_operation.git_init(self.prname)
-            if not self.quiet_on:
-                success_text.msg(self.lang, self.name, self.prname)
-            sys.exit(exit_codes.SUCCESS)
-        else:
-            desc_text.show(self.name, self.version, self.lang)
-            usage_text.show(self.name, self.version, self.lang)