+++ ppi/static/exit_codes.py
@@ -1,9 +1,0 @@
-"""
-exit_codes.py: Standard exit codes.
-Author: Niklas Larsson
-Date: November 19, 2021
-"""
-
-
-ERROR: int = 1
-SUCCESS: int = 0